ML/CV/Unet_Segmentation_2/gradio_projects/requirements.txt
onnxruntime-gpu
numpy
pillow
gradio
opencv-python
torchvision
albumentations
